+++ packages/core/src/providers/index.ts
@@ -1,4 +1,4 @@
-import type { InternalProvider, Profile } from "../types.js"
+import type { Profile } from "../types.js"
 import CredentialsProvider from "./credentials.js"
 import type {
   CredentialsConfig,
@@ -17,25 +17,6 @@ import type { WebAuthnConfig, WebAuthnProviderType } from "./webauthn.js"
 export * from "./credentials.js"
 export * from "./email.js"
 export * from "./oauth.js"
-
-export function isOIDCProvider(
-  provider: InternalProvider<"oidc" | "oauth">
-): provider is InternalProvider<"oidc"> {
-  return provider.type === "oidc"
-}
-
-export function isOAuth2Provider(
-  provider: InternalProvider<"oidc" | "oauth">
-): provider is InternalProvider<"oauth"> {
-  return provider.type === "oauth"
-}
-
-/** Either OAuth 2 or OIDC */
-export function isOAuthProvider(
-  provider: InternalProvider<any>
-): provider is InternalProvider<"oauth" | "oidc"> {
-  return provider.type === "oauth" || provider.type === "oidc"
-}
 
 /**
  * Providers passed to Auth.js must define one of these types.
